requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
appdirs==1.4.4
asgiref==3.8.1
attrs==25.3.0
boto3==1.34.20
botocore==1.34.20
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.4.1
cmake==4.0.0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Django==4.2.7
django-extensions==3.2.3
dlib==19.24.6
docutils==0.20.1
fonttools==4.56.0
frozenlist==1.5.0
funcy==2.0
gensim==4.3.2
idna==3.6
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.5
MarkupSafe==3.0.2
matplotlib==3.8.2
mlxtend==0.23.4
multidict==6.3.2
mysql-connector-python==9.2.0
numexpr==2.10.2
numpy==1.26.2
packaging==23.2
pandas==2.1.4
pillow==11.1.0
propcache==0.3.1
pyLDAvis==3.4.1
PyMySQL==1.1.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
s3transfer==0.10.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
smart-open==6.4.0
sqlparse==0.5.3
stop-words==2018.7.23
threadpoolctl==3.6.0
tqdm==4.66.1
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.0.7
yarl==1.18.3
